backend-service/requirements.txt
fastapi==0.109.2
firebase_admin==6.4.0
openai==1.12.0
pydantic==2.6.1
pymongo==4.6.1
python-dotenv==1.0.1
uvicorn==0.27.1
pytest==8.0.0
stripe==8.3.0
reportlab==4.1.0
pyppeteer==2.0.0
python-multipart==0.0.9